🎨 Reduce surface that requires openmc
